# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Option:
    properties:
      id:
        type: string
      is_correct:
        type: boolean
      text:
        type: string
    type: object
  models.OptionResponse:
    properties:
      id:
        type: string
      text:
        type: string
    type: object
  models.QuestionDifficulty:
    enum:
    - easy
    - medium
    - hard
    type: string
    x-enum-varnames:
    - Easy
    - Medium
    - Hard
  models.QuestionResponse:
    properties:
      content:
        type: string
      created_at:
        type: string
      created_by:
        type: integer
      difficulty:
        $ref: '#/definitions/models.QuestionDifficulty'
      explanation:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      options:
        items:
          $ref: '#/definitions/models.OptionResponse'
        type: array
      points:
        type: integer
      tags:
        items:
          type: string
        type: array
      time_limit:
        type: integer
      title:
        type: string
      type:
        $ref: '#/definitions/models.QuestionType'
      updated_at:
        type: string
    type: object
...
  models.QuestionType:
    enum:
    - multiple_choice
    - true_false
    type: string
    x-enum-varnames:
    - MultipleChoice
    - TrueFalse
...
  services.CreateQuestionRequest:
    properties:
      content:
        type: string
      difficulty:
        $ref: '#/definitions/models.QuestionDifficulty'
      explanation:
        type: string
      options:
        items:
          $ref: '#/definitions/models.Option'
        minItems: 2
        type: array
      points:
        minimum: 1
        type: integer
      tags:
        items:
          type: string
        minItems: 1
        type: array
      time_limit:
        minimum: 10
        type: integer
      title:
        type: string
      type:
        $ref: '#/definitions/models.QuestionType'
    required:
    - content
    - difficulty
    - options
    - tags
    - title
    - type
    type: object
...
  services.QuestionListResponse:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      questions:
        items:
          $ref: '#/definitions/models.QuestionResponse'
        type: array
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  services.UpdateQuestionRequest:
    properties:
      content:
        type: string
      difficulty:
        $ref: '#/definitions/models.QuestionDifficulty'
      explanation:
        type: string
      is_active:
        type: boolean
      options:
        items:
          $ref: '#/definitions/models.Option'
        minItems: 2
        type: array
      points:
        minimum: 1
        type: integer
      tags:
        items:
          type: string
        minItems: 1
        type: array
      time_limit:
        minimum: 10
        type: integer
      title:
        type: string
      type:
        $ref: '#/definitions/models.QuestionType'
    required:
    - content
    - difficulty
    - options
    - tags
    - title
    - type
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com
  description: This is a sample Gin server.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://example.com/terms/
  title: My Gin API
  version: "1.0"
paths:
...
  /api/v1/questions:
    get:
      consumes:
      - application/json
      description: Get a paginated list of questions with optional filtering
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Comma-separated list of tags
        in: query
        name: tags
        type: string
      - description: Question difficulty (easy, medium, hard)
        in: query
        name: difficulty
        type: string
      - description: Question type (multiple_choice, true_false)
        in: query
        name: type
        type: string
      - description: Search term
        in: query
        name: search
        type: string
      - description: Filter by active status
        in: query
        name: is_active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Questions list
          schema:
            $ref: '#/definitions/services.QuestionListResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get questions list
      tags:
      - questions
    post:
      consumes:
      - application/json
      description: Create a new question (admin only)
      parameters:
      - description: Question data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.CreateQuestionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Question created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create question
      tags:
      - questions
  /api/v1/questions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific question (admin only)
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Question deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Question not found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Question is used in active exams
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete question
      tags:
      - questions
    get:
      consumes:
      - application/json
      description: Get a specific question by its ID
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Question details
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Question not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get question by ID
      tags:
      - questions
    put:
      consumes:
      - application/json
      description: Update a specific question (admin only)
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: integer
      - description: Question update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.UpdateQuestionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Question updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Question not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update question
      tags:
      - questions
  /api/v1/questions/random:
    get:
      consumes:
      - application/json
      description: Get random questions filtered by tags and difficulty level
      parameters:
      - description: Comma-separated list of tags
        in: query
        name: tags
        type: string
      - default: 10
        description: Number of questions to return
        in: query
        name: count
        type: integer
      - description: Question difficulty (easy, medium, hard)
        in: query
        name: difficulty
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Random questions
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get random questions by tags
      tags:
      - questions
  /api/v1/questions/tags:
    get:
      consumes:
      - application/json
      description: Get all available question tags
      produces:
      - application/json
      responses:
        "200":
          description: Tags list
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get question tags
      tags:
      - questions
...
swagger: "2.0"